fix: address supervisor review findings for chunk A
- Delete domains/contribution/ shim per prelude (outright deletion, not shim)
- Remove check_sync_state/write_sync_state stubs from distribution/state.py
- Fix git-commit-failure branch in warehouse/contribute.py (now raises)
- Add pruned_paths to SyncSummary; orchestrator no longer references missing field
- Replace anonymous type("Orphan", ...) with proper Orphan dataclass
- Extract duplicated _tracked_paths helpers to warehouse/_tracked_paths.py
- Move inline imports to module top (project-standard compliance)
- Delete dead expand_artifact_paths method
- Simplify dead elif branch in classify_entries
- Collapse identical branches in migration.py contribute path
