--- libs/beacon/src/beacon/domains/contribution/__init__.py
@@ -1,1 +1,0 @@
-"""Shim for backward compatibility during transition."""
