+++ libs/beacon/src/beacon/domains/warehouse/contribute.py
@@ -10,8 +10,8 @@ from pathlib import Path
 
 from loguru import logger
 
-from beacon.core.manifest.beacon import BeaconManifest
 from beacon.core.preconditions import ensure_sync_ready
+from beacon.domains.warehouse._tracked_paths import _get_tracked_paths
 
 
 @dataclass
@@ -33,53 +33,6 @@ def _run_git(
         text=True,
         timeout=timeout,
     )
-
-
-def _get_tracked_paths(warehouse_path: Path, beacon_yaml: Path) -> list[str]:
-    """Return the list of beacon.yaml-matched paths relative to warehouse root."""
-    if not beacon_yaml.exists():
-        return []
-
-    beacon_settings = BeaconManifest.from_yaml(beacon_yaml)
-    paths: list[str] = []
-
-    for pattern in beacon_settings.artifacts.knowledge:
-        paths.extend(_expand_pattern(warehouse_path, pattern))
-
-    for pattern in beacon_settings.artifacts.skills:
-        paths.extend(_expand_pattern(warehouse_path, pattern))
-
-    for pattern in beacon_settings.artifacts.contexts:
-        paths.extend(_expand_pattern(warehouse_path, pattern))
-
-    return paths
-
-
-def _expand_pattern(warehouse_path: Path, pattern: str) -> list[str]:
-    """Expand a beacon.yaml pattern to concrete relative paths."""
-    import glob
-
-    if "*" in pattern or "?" in pattern:
-        matches = glob.glob(str(warehouse_path / pattern), recursive=True)
-        return [
-            str(Path(m).relative_to(warehouse_path))
-            for m in matches
-            if Path(m).is_file()
-        ]
-
-    p = warehouse_path / pattern
-    if p.is_dir():
-        matches = glob.glob(str(p / "**" / "*"), recursive=True)
-        return [
-            str(Path(m).relative_to(warehouse_path))
-            for m in matches
-            if Path(m).is_file()
-        ]
-
-    if p.is_file():
-        return [pattern]
-
-    return [pattern]
 
 
 def contribute(
@@ -126,9 +79,8 @@ def contribute(
     commit_result = _run_git(warehouse_path, ["commit", "-m", message])
     if commit_result.returncode != 0:
         logger.error("Git commit failed: {}", commit_result.stderr)
-        return ContributeResult(
-            status="no_changes",
-            message=commit_result.stderr,
+        raise RuntimeError(
+            f"Git commit failed in warehouse: {commit_result.stderr.strip()}"
         )
 
     # Get the committed SHA
