+++ libs/beacon/src/beacon/domains/warehouse/status.py
@@ -8,8 +8,8 @@ import subprocess
 from dataclasses import dataclass, field
 from pathlib import Path
 
-from beacon.core.manifest.beacon import BeaconManifest
 from beacon.core.preconditions import ensure_sync_ready
+from beacon.domains.warehouse._tracked_paths import _get_tracked_paths
 
 
 @dataclass
@@ -41,54 +41,6 @@ def _run_git(
         text=True,
         timeout=timeout,
     )
-
-
-def _get_tracked_paths(warehouse_path: Path, beacon_yaml: Path) -> list[str]:
-    """Return the list of beacon.yaml-matched paths relative to warehouse root."""
-
-    if not beacon_yaml.exists():
-        return []
-
-    beacon_settings = BeaconManifest.from_yaml(beacon_yaml)
-    paths: list[str] = []
-
-    for pattern in beacon_settings.artifacts.knowledge:
-        paths.extend(_expand_pattern(warehouse_path, pattern))
-
-    for pattern in beacon_settings.artifacts.skills:
-        paths.extend(_expand_pattern(warehouse_path, pattern))
-
-    for pattern in beacon_settings.artifacts.contexts:
-        paths.extend(_expand_pattern(warehouse_path, pattern))
-
-    return paths
-
-
-def _expand_pattern(warehouse_path: Path, pattern: str) -> list[str]:
-    """Expand a beacon.yaml pattern to concrete relative paths."""
-    import glob
-
-    if "*" in pattern or "?" in pattern:
-        matches = glob.glob(str(warehouse_path / pattern), recursive=True)
-        return [
-            str(Path(m).relative_to(warehouse_path))
-            for m in matches
-            if Path(m).is_file()
-        ]
-
-    p = warehouse_path / pattern
-    if p.is_dir():
-        matches = glob.glob(str(p / "**" / "*"), recursive=True)
-        return [
-            str(Path(m).relative_to(warehouse_path))
-            for m in matches
-            if Path(m).is_file()
-        ]
-
-    if p.is_file():
-        return [pattern]
-
-    return [pattern]
 
 
 def status(
